Modify html and css, add style to first recipe
Some styling parameters have been modified for better aesthetics.
Also, lots of new styling was added to display the recipe pages as
intended, and linked to the first recipe. The html file forthe first
recipe has been modified to display images under each step correctly,
instead of the previous way of doing it, which was sketchy.